task2_conversational_bot/requirements.txt
langchain==0.0.353
openai==1.12.0
python-dotenv
faiss-cpu
pypdf
tiktoken
